requirements.txt
Django
pytz
psycopg2
psycopg2-binary
redis
coverage
pytest
pytest-cov
pytest-django
codacy-coverage
sorl-thumbnail
Pillow
django-phonenumber-field
phonenumbers
whitenoise
django-widget-tweaks
djangorestframework
django-bootstrap4
django-heroku
django-choices
gunicorn
django-tables2
tablib
django-filter
django-debug-toolbar
django-model-utils